+++ kzg10-rust/kzg-bench/src/lib.rs
@@ -1,1 +1,5 @@
+
+#[macro_escape] mod test_macro;
 mod test;
+mod poly;
+mod fft_common;
