+++ kzg10-rust/kzg-bench/src/test.rs
@@ -17,133 +17,6 @@ use mcl_rust::g1::G1;
 use mcl_rust::g2::G2;
 use mcl_rust::gt::GT;
 
-macro_rules! field_test {
-    ($t:ty) => {{
-        let mut x = <$t>::zero();
-        assert!(x.is_valid());
-        assert!(x.is_zero());
-        assert!(!x.is_one());
-        x.set_int(1);
-        assert!(!x.is_zero());
-        assert!(x.is_one());
-        let mut y = <$t>::from_int(1);
-        assert!(y.is_valid());
-        assert_eq!(x, y);
-        y.set_int(2);
-        assert!(x != y);
-        x.set_str("65535", 10);
-        y.set_str("ffff", 16);
-        assert!(x.is_valid());
-        assert_eq!(x, y);
-        x.set_int(123);
-        assert!(x.is_odd());
-        x.set_int(124);
-        assert!(!x.is_odd());
-        assert!(!x.is_negative());
-        x.set_int(-124);
-        assert!(x.is_negative());
-
-        let mut z = unsafe { <$t>::uninit() };
-        let mut w = unsafe { <$t>::uninit() };
-
-        let a = 256;
-        let b = 8;
-        x.set_int(a);
-        y.set_int(b);
-        <$t>::add(&mut z, &x, &y);
-        w.set_int(a + b);
-        assert_eq!(z, w);
-        assert_eq!(w, (&x + &y));
-        z = x.clone();
-        z += &y;
-        assert_eq!(z, w);
-
-        <$t>::sub(&mut z, &x, &y);
-        w.set_int(a - b);
-        assert_eq!(z, w);
-        assert_eq!(w, (&x - &y));
-        z = x.clone();
-        z -= &y;
-        assert_eq!(z, w);
-
-        <$t>::mul(&mut z, &x, &y);
-        w.set_int(a * b);
-        assert_eq!(z, w);
-        assert_eq!(w, (&x * &y));
-        z = x.clone();
-        z *= &y;
-        assert_eq!(z, w);
-
-        <$t>::div(&mut z, &x, &y);
-        w.set_int(a / b);
-        assert_eq!(z, w);
-        assert_eq!(z, (&x / &y));
-        z = x.clone();
-        z /= &y;
-        assert_eq!(z, w);
-
-        assert!(x.set_little_endian_mod(&[1, 2, 3, 4, 5]));
-        assert_eq!(x.get_str(16), "504030201");
-        <$t>::sqr(&mut y, &x);
-        <$t>::mul(&mut z, &x, &x);
-        assert_eq!(y, z);
-
-        assert!(<$t>::square_root(&mut w, &y));
-        if w != x {
-            <$t>::neg(&mut z, &w);
-            assert_eq!(x, z);
-        }
-    }};
-}
-
-macro_rules! ec_test {
-    ($t:ty, $f:ty, $P:expr) => {
-        #[allow(non_snake_case)]
-        assert!($P.is_valid());
-        assert!(!$P.is_zero());
-        let mut P1 = <$t>::zero();
-        assert!(P1.is_zero());
-        assert_ne!(P1, $P);
-        <$t>::neg(&mut P1, &$P);
-        let mut x: $f = unsafe { <$f>::uninit() };
-        <$f>::neg(&mut x, &P1.y);
-        assert_eq!(&x, &$P.y);
-
-        <$t>::dbl(&mut P1, &$P);
-        let mut P2: $t = unsafe { <$t>::uninit() };
-        let mut P3: $t = unsafe { <$t>::uninit() };
-        <$t>::add(&mut P2, &$P, &$P);
-        assert_eq!(P2, P1);
-        <$t>::add(&mut P3, &P2, &$P);
-        assert_eq!(P3, (&P2 + &$P));
-        assert_eq!(P2, (&P3 - &$P));
-        let mut y: Fr = Fr::from_int(1);
-        <$t>::mul(&mut P2, &$P, &y);
-        assert_eq!(P2, $P);
-        y.set_int(2);
-        <$t>::mul(&mut P2, &$P, &y);
-        assert_eq!(P2, P1);
-        y.set_int(3);
-        <$t>::mul(&mut P2, &$P, &y);
-        assert_eq!(P2, P3);
-        P2 = P1.clone();
-        P2 += &$P;
-        assert_eq!(P2, P3);
-
-        P2 -= &$P;
-        assert_eq!(P2, P1);
-    };
-}
-
-macro_rules! serialize_test {
-    ($t:ty, $x:expr) => {
-        let buf = $x.serialize();
-        let mut y: $t = unsafe { <$t>::uninit() };
-        assert!(y.deserialize(&buf));
-        assert_eq!($x, y);
-    };
-}
-
 #[test]
 #[allow(non_snake_case)]
 fn mcl_test() {
@@ -372,29 +245,6 @@ fn proof_loop_works_with_random_points() {
 }
 
 // FFT | FK20
-
-#[test]
-fn fftsettings_new_creates_valid_settings() {
-    // Arrange
-    assert!(init(CurveType::BLS12_381));
-    // Act
-    let fft_settings = FFTSettings::new(4 + 5 + 1);
-    
-    // Assert
-    let expected = vec![
-        Fr::from_int(1),
-        Fr::from_str("21328829733576761151404230261968752855781179864716879432436835449516750606329", 10).unwrap(),
-        Fr::from_str("12531186154666751577774347439625638674013361494693625348921624593362229945844", 10).unwrap(),
-        Fr::from_str("36815421669481109810171413925233110915304823983913164224028689762034127238951", 10).unwrap()
-    ];
-
-    // has lots more members, but checking the first few should point out whether the math is correct
-    let starting_equal =fft_settings.exp_roots_of_unity.iter()
-        .zip(expected)
-        .all(|(a, b)| a.get_str(10) == b.get_str(10));
-
-    assert!(starting_equal);
-}
 
 #[test]
 fn fk20matrix_new_builds_valid_settings() {
@@ -950,104 +800,6 @@ fn polynomial_recover_from_samples_should_recover_all_values_given_less_than_hal
     assert_eq!(data.len(), recovered.order());
     for i in 0..data.len() {
         assert_eq!(data[i].get_str(10), recovered.coeffs[i].get_str(10));
-    }
-}
-
-#[test]
-fn root_of_unity_is_the_expected_size() {
-    //Arrange
-    let element_size: usize;
-    let length: usize;
-    unsafe{
-        element_size = mem::size_of_val(&SCALE_2_ROOT_OF_UNITY[0]);
-        length = SCALE_2_ROOT_OF_UNITY.len()
-    }
-    
-    //Act/Assert
-    assert_eq!(mem::size_of::<Fr>(), element_size);
-    assert_eq!(32, length);
-}
-
-#[test]
-// #[should_panic]
-fn roots_of_unity_out_of_bounds_fails() {
-    //Arrange
-    assert!(init(CurveType::BLS12_381));
-
-    //Act
-    let settings = std::panic::catch_unwind(|| FFTSettings::new(32));
-    
-    //Assert
-    assert!(settings.is_err());
-}
-
-#[test]
-fn roots_of_unity_are_plausible() {
-    //Arrange
-    assert!(init(CurveType::BLS12_381));
-    let mut i = 0;
-    let mut r: Fr;
-
-    //Act/Assert
-    while i < 32 {
-        unsafe{
-            r = SCALE_2_ROOT_OF_UNITY[i].clone();
-        }
-        
-        let mut j = 0;
-        while j < i {
-            let cl = r.clone();
-            Fr::sqr(&mut r, &cl);
-            j = j + 1;
-        }
-        assert!(r.is_one());
-        i = i + 1;
-    }
-}
-
-#[test]
-fn expand_roots_is_plausible() {
-    //Arrange
-    const SCALE: usize = 15;
-    const WIDTH: usize = 1 << SCALE;
-    let root: Fr;
-    let mut prod: Fr = Fr::zero();
-
-    //Act/Assert
-    unsafe{
-        root = SCALE_2_ROOT_OF_UNITY[SCALE].clone();
-    }
-    let expanded = expand_root_of_unity(&root);
-    assert!(expanded[0].is_one());
-    assert!(expanded[WIDTH].is_one());
-
-    for i in 1..(WIDTH/2 + 1) {
-        Fr::mul(&mut prod, &expanded[i], &expanded[WIDTH - i]);
-        assert!(prod.is_one());
-    }
-
-}
-
-#[test]
-fn poly_test_div() {
-    let first: Vec<i32> = vec![-1, 0, 1];
-    let second: Vec<Fr> = vec![Fr::from_int(1), Fr::from_int(1)];
-    let third: Vec<i32> = vec![-1, 1];
-
-    let dividend = Polynomial::from_i32(&first);
-    let expected = Polynomial::from_i32(&third);
-
-    let result = dividend.long_division(&second);
-
-    let expected_len: usize;
-    let result_len: usize;
-
-    expected_len = expected.coeffs.len();
-    result_len = result.coeffs.len();
-
-    assert_eq!(expected_len, result_len);
-    for i in 1..result_len {
-     assert_eq!(result.coeffs[i], expected.coeffs[i]);
     }
 }
